git-http-backend: Linted code
