+++ types/git-http-backend/index.d.ts
@@ -1,13 +1,35 @@
 /// <reference types='node' />
-
-declare module "git-http-backend" {
-    export default Backend;
-    import { Duplex } from "stream";
-    
-    class Backend extends Duplex {
-        constructor(uri: string, cb: (err: any, service: import("./service")) => void);
-        parsed: boolean;
-        service: string | string[];
-        info: boolean;
-    }
+import { Duplex, Writable } from "stream";
+interface Fields {
+    head: string;
+    last: string;
+    refname: string;
+    ref: string;
+    tag: string;
+    branch: string;
+    name: string;
 }
+declare class Service {
+    constructor(opts: {
+        info: boolean,
+        cmd: string,
+    } & Fields, backend: Backend);
+    info: boolean;
+    cmd: string;
+    _bands: Buffer[];
+    action: string | {"git-receive-pack": string, "git-upload-pack": "pull"};
+    type: string;
+    _backend: Backend;
+    fields: Fields;
+    args: string[];
+    createStream(): Duplex;
+    createBand(): Writable;
+}
+
+declare class Backend extends Duplex {
+    constructor(uri: string, cb: (err: any, service: Service) => void);
+    parsed: boolean;
+    service: string | string[];
+    info: boolean;
+}
+export default Backend;
