+++ types/git-http-backend/tslint.json
@@ -1,1 +1,1 @@
-{ "extends": "@definitelytyped/dtslint/dt.json" }
+{ "extends": "@definitelytyped/dtslint/dtslint.json" }
